Fix bug in loadGroupMembers manual override
The captains@codefortulsa.org email triggers an error in the
group.hasUser function for some reason, and I added this manual override
so it wouldn't spam the logs. But the manual override didn't work, and
now I see that it was very broken:

- Most emails (that aren't manually overridden) get a -1 back from
indexOf, resulting in `false` when not'd. The `if` statement falls
through to the else and assumes that the user is subscribed to the list.
- The email that was manually overridden gets a 0 back from indexOf,
since it is the only email in the array, and then that gets not'd to be
`true`, which is exactly the opposite of the intent of this check.

So in effect, only the first email in the MANUAL_OVERRIDE_ADD_MEMBER
array was actually being checked for membership. :flushed: