Improve logic for Donation name reporting
For donations from organizations (corporations, governments, and
foundations), we want to report them as:

  SomeCorp (Sally Jones)

rather than just from "Sally Jones". This will allow the brigades to
remember who their point of contact there was.

Note that having a Contact associated with an organization donation is
somewhat rare -- typically this happens when the organization uses the
public donate form.