Add WIP MeetupToSalesforce integration
This got git-cleaned up from another branch so I'm adding the WIP code
here.

This code uses an approach of a "staging" Google Doc that is used to
persist lists of contacts and affiliations for manual inspection. It's
likely that eventually we can get rid of this doc and just do everything
in-memory.