Add rowsToCSV helper method
This method converts a 2d array of rows into a CSV. Pretty simple if you
don't use it with data that has quotation marks or commas within.

This method is smart to some data formatting, starting with Date
objects. ISO 8601 is supported by Salesforce, so we'll be opinionated
and use that format.