Implement meetupToSalesforceExecute
I had left this as an exercise to the reader, but now it's implemented
in terms of the three step process outlined in the comments within.

Also:
- Converts EXISTING_AFFILIATIONS from a beginning-of-script cache to
just be a callable method, because between the prepare and execute
steps, ideally we'd refresh this cache.

- Fixes the relationship name in the CSV headers:
`npe5__Contact__r:Meetup_User_ID__c` => `npe5__Contact__r.Meetup_User_ID__c`