Refactor Meetup-related classes and add tests
For better testability, as well as to take advantage of the compilation
pipeline, this separates the Meetup functionality into a few classes
that can be more easily tested.

Thus begins the herculean effort of implementing mocks for all of the
GAS runtime... but for now it's just UrlFetchApp and the
PropertiesService.