Update rowsToCSV to actually escape and quote properly
When using real world data, it became evident that it's time to write a
CSV generator that is a bit more capable of handling creatively
formatted data. For example, I had assumed names wouldn't include
quotation marks but a lot of users on Meetup have their "nicknames" in
their name.