Run Jasmine tests with @babel/node
Since jasmine tests run with the nodejs runtime, they cannot support
the fancy syntax of Typescript / ES6 modules without help from Babel.

This commit extracts the babel config from webpack into a shared
`.babelrc` file, adds the magical `@babel/env` preset that just kinda
bundles everything in, and tweaks the test suite to work in a strict
module environment (the value of `this` is undefined at the top-level,
so all tests need to be enclosed in a non-arrow-`function() {`).